+++ scripts/macros/2024/classFeatures/paladin/paladinsSmite.js
@@ -1,5 +1,5 @@
 export let paladinsSmite = {
     name: 'Paladin\'s Smite',
-    version: '1.3.148',
+    version: '1.5.20',
     rules: 'modern'
 };